a function that prints integer

+++ 2-mandatory.c
@@ -7,8 +7,8 @@
 int print_integer(va_list pa)
 {
 	int len = 0, i = 0, j = 0;
-	int num = va_arg(pa, int);
-	int number = num, temp;
+	long int num = va_arg(pa, int);
+	long int number = num, temp;
 	char *ptr;
 
 	if (number < 0)
@@ -49,8 +49,8 @@ int print_integer(va_list pa)
 int print_double(va_list pa)
 {
 	int len = 0, i = 0, j = 0;
-	int num = va_arg(pa, int);
-	int number = num, temp;
+	long int num = va_arg(pa, int);
+	 long int number = num, temp;
 	char *ptr;
 
 	if (number < 0)
